Master/requirements.txt
django-cors-headers==3.14.0
djangorestframework==3.14.0
Django==4.2